Day13/Functions/requirements.txt
absl-py==0.8.1
astor==0.8.1
azure-functions==1.0.3
Click==7.0
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.17.4
opt-einsum==3.1.0
protobuf==3.11.1
PyYAML==5.2
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
textgenrnn==1.5.0
tqdm==4.40.2
Werkzeug==0.16.0
wrapt==1.11.2
